+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Improvements
 
 - [\#339](https://github.com/arkworks-rs/algebra/pull/339) (ark-ff) Remove duplicated code from `test_field` module and replace its usage with `ark-test-curves` crate.
+- [\#354](https://github.com/arkworks-rs/algebra/pull/354) (test-curves) Add benchmarks to `test-curves`.
 
 ### Bugfixes